move files around and rename according to naming convention
